Further isolate entities implementation
