+++ pyxform/entities/entity_declaration.py
